ьфmade load factor float

+++ src/main/java/core/basesyntax/MyHashMap.java
@@ -4,7 +4,7 @@ import java.util.Objects;
 
 public class MyHashMap<K, V> implements MyMap<K, V> {
     private static final int CAPACITY = 16;
-    private static final double LOAD_FACTOR = 0.75;
+    private static final float LOAD_FACTOR = 0.75f;
     private int size;
     private Node<K, V>[] keysValue;
